Version 2.15.9
Changelog for i40e-2.15.9
===========================================================================
Add forwarding GRE UDP and IP packets to VF
Add a check to see if MFS is set
Add warning when adding mirror vlan rule
Add Asym_Pause to supported link modes
Del rx_fragments and rx_jabber from rx_errors
Fix delay after global reset
Fix ATR queue selection
Fix DCB enable flag while TCs are created
Fix the link loss on driver unload
Fix acquire VSI pointer only after VF is initialized
Fix not blinking X722 with x557 PHY via ‘ethtool -p'
Fix add tc filter for IPv6
Fix race condition while deleting MAC/VLAN filters
Fix setting PF MAC filters when changing MAC address
Fix VF notification when changing vlan strip status through sysfs
Fix PHY type identifiers for 2.5G and 5G adapters
Fix display error code in dmesg
Fix display statistics for veb_tc
Fix return ENOPKG when right DDP package is missing
Fix endianness conversions
Fix build failures on current upstream and OSV kernels

NOTE:  Receive Error counts may be higher than the actual packet error count
---------------------------------------------------------------------
When a packet is received with more than one error, two bad packets may be
reported. This affects all devices based on 10G, or faster, controllers.

+++ src/i40e_nvm.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: GPL-2.0
-/* Copyright(c) 2013 - 2020 Intel Corporation. */
+/* Copyright(c) 2013 - 2021 Intel Corporation. */
 
 #include "i40e_prototype.h"
 
@@ -73,7 +73,8 @@ i40e_status i40e_acquire_nvm(struct i40e_hw *hw,
 	if (ret_code)
 		i40e_debug(hw, I40E_DEBUG_NVM,
 			   "NVM acquire type %d failed time_left=%llu ret=%d aq_err=%d\n",
-			   access, time_left, ret_code, hw->aq.asq_last_status);
+			   access, (unsigned long long)time_left, ret_code,
+			   hw->aq.asq_last_status);
 
 	if (ret_code && time_left) {
 		/* Poll until the current NVM owner timeouts */
@@ -95,7 +96,8 @@ i40e_status i40e_acquire_nvm(struct i40e_hw *hw,
 			hw->nvm.hw_semaphore_timeout = 0;
 			i40e_debug(hw, I40E_DEBUG_NVM,
 				   "NVM acquire timed out, wait %llu ms before trying again. status=%d aq_err=%d\n",
-				   time_left, ret_code, hw->aq.asq_last_status);
+				   (unsigned long long)time_left, ret_code,
+				   hw->aq.asq_last_status);
 		}
 	}
